move functions to magpie for UI usage
